Refactor GraphicsSet to graphics_set/GraphicSet
