+++ foundry/core/graphics_set/GraphicsSet.py
