+++ foundry/core/player_animations/PlayerAnimationData.py
@@ -1,8 +1,8 @@
 from attr import attrs
 
+from foundry.core.graphics_set.GraphicsSet import GraphicsSetProtocol
 from foundry.core.palette.Palette import PaletteGroupProtocol
 from foundry.core.player_animations.PlayerAnimation import PlayerAnimation
-from foundry.game.gfx.GraphicsSet import GraphicsSetProtocol
 
 
 @attrs(slots=True, auto_attribs=True)
