+++ foundry/game/gfx/drawable/Sprite.py
@@ -3,11 +3,11 @@ from functools import lru_cache
 from PySide6.QtCore import QPoint
 from PySide6.QtGui import QColor, QImage, QPainter, Qt
 
+from foundry.core.graphics_set.GraphicsSet import GraphicsSetProtocol
 from foundry.core.palette.Palette import NESPalette, PaletteGroupProtocol
 from foundry.game.File import ROM
 from foundry.game.gfx.drawable import MASK_COLOR, apply_selection_overlay
 from foundry.game.gfx.drawable.Tile import Tile
-from foundry.game.gfx.GraphicsSet import GraphicsSetProtocol
 
 
 @lru_cache(2 ** 10)
